requirements.txt
beautifulsoup4>=4.12.3
feedparser>=6.0.11
anthropic>=0.67.0
psycopg[binary]>=3.2.1
python-dotenv>=1.0.1
requests>=2.32.3
Pillow>=11.3.0
google-genai>=1.0.0
discord.py>=2.5.0
jieba>=0.42.1
